Update to newer mongoid index syntax on User model
This change is just for getting the new mongoid to work.  If anyone wants to change this index type to make it better, then feel free.  This change also makes it possible to run a rails server, but opening rubeque in the browser yields a SessionRestoreError. Methinks I'm about to learn quite a bit about devise.

Signed-off-by: Michaux Kelley <michaux.kelley@scimedsolutions.com>